requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
langchain-neo4j>=0.0.5
langgraph>=0.0.40
supabase>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
openai>=1.0.0
neo4j>=5.0.0
typing-extensions>=4.0.0
psutil>=5.9.0 